refactor: move quality-gate to app/evaluation.py, delete app/claude.py
Also routes app/git.py (create_branch, commit/PR) through injected providers
— a call site the plan missed.
