+++ plan.md
@@ -85,7 +85,7 @@ A `LLMProvider` Protocol with one async method (`run`) and three implementations
 
 - [x] Now that the per-stage sites use providers, inject the decomposer's provider. In `app/agents.py::decompose_task`, accept an injected `provider` parameter (the Decomposition stage's provider) instead of importing `call_claude`, and update its three callers in `app/runner.py` (≈lines 121, 219, 384) to pass `stage_map["Decomposition"]`'s provider. Leave `app/claude.py` in place for now — it is still used by `evaluate_should_iterate`. Run `uv run pytest` and confirm a no-config run still decomposes via Claude.
 
-- [ ] Now that decompose uses an injected provider, move the quality-gate off `call_claude` and delete the old module. Move `evaluate_should_iterate` from `app/claude.py` into a new `app/evaluation.py`, taking an injected `provider` parameter and recording its own cost via `pipeline_stats.add_call`. Update its callers at `app/runner.py:182` and `app/runner_steps.py:215` to pass a provider: pass `ClaudeProvider()` for now — the `--config` task replaces it with the resolved `[defaults]` provider. Delete `app/claude.py`. Run `uv run pytest` and `uv run python -c "import app.runner, app.pipeline, app.workers, app.agents, app.evaluation"` to confirm no dangling `app.claude` import remains.
+- [x] Now that decompose uses an injected provider, move the quality-gate off `call_claude` and delete the old module. Move `evaluate_should_iterate` from `app/claude.py` into a new `app/evaluation.py`, taking an injected `provider` parameter and recording its own cost via `pipeline_stats.add_call`. Update its callers at `app/runner.py:182` and `app/runner_steps.py:215` to pass a provider: pass `ClaudeProvider()` for now — the `--config` task replaces it with the resolved `[defaults]` provider. Delete `app/claude.py`. Run `uv run pytest` and `uv run python -c "import app.runner, app.pipeline, app.workers, app.agents, app.evaluation"` to confirm no dangling `app.claude` import remains.
 
 - [ ] Now that config and providers are wired, add binary preflight. In `app/config.py` add `preflight(resolved) -> list[str]` returning install-instruction messages for any referenced provider whose binary is missing (`shutil.which("claude"|"codex"|"gemini")`), checking only providers present in `resolved`. Call it in `app/runner.py` right after `pipeline_stats.reset()` (≈line 44) and abort the run with the messages written to the log if any are missing. Write `tests/test_config.py::test_preflight_reports_missing` monkeypatching `shutil.which`. Run `uv run pytest tests/test_config.py`.
